# contrato-api/openapi.yaml
openapi: 3.1.0
info:
  title: ValidaLab API
  summary: Contrato de API único de ValidaLab — fuente de verdad entre frontend y backend.
  description: |
    Este documento OpenAPI es la **fuente de verdad única** del contrato de API de ValidaLab.
    Tanto `frontend/` como `backend/` desarrollan contra este documento sin conocer ni depender
    del código del otro paquete. Cualquier endpoint, payload o comportamiento de borde entre
    cliente y servidor se define **aquí primero** y solo después se implementa.

    Es un **único documento** (no fragmentado por módulos). La navegación por dominio se hace
    mediante los `tags` declarados más abajo (`usuarios`, `ideas`, `contactos`, `entrevistas`,
    `kpis`, `agente`, `proveedores`), que se corresponden con los módulos de dominio del backend.

    ## Convenciones transversales

    Estas reglas aplican a **todos** los endpoints. El resumen normativo vive también en el
    `CLAUDE.md` de la raíz; el detalle (catálogo de errores, esquemas concretos) vive aquí.

    ### Autenticación
    Todas las operaciones (salvo registro y login) requieren un **JWT** en
    `Authorization: Bearer <token>` (esquema `bearerAuth`). Una petición sin token válido
    devuelve `401` con el error `NO_AUTENTICADO`.

    ### Aislamiento multi-tenant por `owner_id` (RNF-04, RNF-05)
    ValidaLab es multi-tenant desde la primera versión. El `owner_id` se deriva **siempre** del
...

    ### Formato de error
...
    en este documento, no en `CLAUDE.md`.

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /ideas:
    post:
      tags: [ideas]
      summary: Crear una idea
      description: Crea una idea en estado `borrador`, asociada al usuario autenticado (`ownerId` derivado del token).
      operationId: crearIdea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearIdeaRequest'
      responses:
        '201':
          description: Idea creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
    get:
      tags: [ideas]
      summary: Listar las ideas propias
      description: Devuelve solo las ideas del usuario autenticado, paginadas. Admite filtro opcional por `estado`.
      operationId: listarIdeas
      parameters:
        - $ref: '#/components/parameters/pagina'
        - $ref: '#/components/parameters/porPagina'
        - $ref: '#/components/parameters/filtroEstadoIdea'
      responses:
        '200':
          description: Página de ideas propias.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdeasPaginadas'
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /ideas/{id}:
    get:
      tags: [ideas]
      summary: Consultar una idea propia
      operationId: obtenerIdea
      parameters:
        - $ref: '#/components/parameters/idIdea'
      responses:
        '200':
          description: Idea solicitada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
    patch:
      tags: [ideas]
      summary: Editar el contenido de una idea propia
      description: Edita el contenido (`titulo`, `descripcion`, `problema`, `segmentoBeachhead`). No permite fijar el `estado` a `go`/`pivote`/`kill`, que provienen del veredicto aprobado (E6).
      operationId: actualizarIdea
      parameters:
        - $ref: '#/components/parameters/idIdea'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarIdeaRequest'
      responses:
        '200':
          description: Idea actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
  /ideas/{id}/archivar:
    post:
      tags: [ideas]
      summary: Archivar una idea
      description: Marca la idea propia como `archivada` conservando su evidencia; no la elimina.
      operationId: archivarIdea
      parameters:
        - $ref: '#/components/parameters/idIdea'
      responses:
        '200':
          description: Idea archivada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
  /ideas/{id}/desarchivar:
    post:
      tags: [ideas]
      summary: Reabrir (desarchivar) una idea archivada
      description: Devuelve una idea `archivada` al estado `borrador` para retomarla, conservando su evidencia. Si la idea no estaba archivada, responde `409`.
      operationId: desarchivarIdea
      parameters:
        - $ref: '#/components/parameters/idIdea'
      responses:
        '200':
          description: Idea reabierta en estado `borrador`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          $ref: '#/components/responses/Conflicto'

  # --- Hipótesis y umbrales por idea (épica E2), anidados bajo el tag `ideas` ---
...
  /guiones:
    post:
      tags: [entrevistas]
      summary: Crear un guión de entrevista
      description: |
        Crea un guión reutilizable **entre ideas** (RF-08, HU-11), con sus preguntas ordenadas.
        El `ownerId` se deriva del token, nunca del cuerpo.
      operationId: crearGuion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearGuionRequest'
      responses:
        '201':
          description: Guión creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guion'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
    get:
      tags: [entrevistas]
      summary: Listar los guiones propios
      description: Devuelve solo los guiones del usuario autenticado, paginados.
      operationId: listarGuiones
      parameters:
        - $ref: '#/components/parameters/pagina'
        - $ref: '#/components/parameters/porPagina'
      responses:
        '200':
          description: Página de guiones propios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuionesPaginados'
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /guiones/{idGuion}:
    get:
      tags: [entrevistas]
      summary: Consultar un guión propio
      operationId: obtenerGuion
      parameters:
        - $ref: '#/components/parameters/idGuion'
      responses:
        '200':
          description: Guión solicitado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guion'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
    patch:
      tags: [entrevistas]
      summary: Editar un guión propio
      description: |
        Edita el `nombre`, la `descripcion` y/o las `preguntas` (reemplazando el conjunto ordenado).

        **Un guión con evidencia es inmutable en su estructura.** Si el guión ya está referenciado
        por al menos una entrevista y el cuerpo incluye `preguntas`, responde `409 CONFLICTO` y las
        preguntas permanecen intactas. No es una restricción burocrática: cada `RespuestaEntrevista`
        apunta a un `preguntaId` de este guión, y como `PreguntaRequest` no lleva `id`, un reemplazo
        rompería esa referencia. Y aunque se conservara el identificador, cambiar el **texto** de una
        pregunta ya respondida haría que la respuesta contestara a otra cosa, corrompiendo en silencio
        la evidencia sobre la que se calculan scores, KPIs y veredicto (RNF-15).

        El `nombre` y la `descripcion` **siguen siendo editables** aunque haya entrevistas: no
        participan en la evidencia. Un guión sin entrevistas se edita sin restricción alguna.
      operationId: actualizarGuion
      parameters:
        - $ref: '#/components/parameters/idGuion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarGuionRequest'
      responses:
        '200':
          description: Guión actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guion'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          $ref: '#/components/responses/Conflicto'
        '422':
          $ref: '#/components/responses/ValidacionFallida'
    delete:
      tags: [entrevistas]
      summary: Eliminar un guión propio
      description: |
        Elimina un guión propio **que ninguna entrevista referencie**.

        Si al menos una entrevista lo usa, responde `409 CONFLICTO` y el guión permanece: borrarlo
        dejaría colgando tanto su `guionId` como los `preguntaId` de todas las respuestas ya
        registradas, y esas entrevistas son la evidencia desde la que deben poder reconstruirse los
        KPIs (RNF-15). Es la misma postura que el contrato ya sostiene al permitir **archivar** una
        idea en lugar de eliminarla.
      operationId: eliminarGuion
      parameters:
        - $ref: '#/components/parameters/idGuion'
      responses:
        '204':
          description: Guión eliminado; sin contenido.
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          $ref: '#/components/responses/Conflicto'

  # --- Entrevistas y scoring por IA (épica E4), anidadas bajo la idea ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por ValidaLab tras el login. El `owner_id` del usuario se deriva del token;
        nunca se acepta como entrada del cliente.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: |
        Refresh token opaco transportado en una cookie `HttpOnly; Secure; SameSite=Strict` con
        `Path` acotado a las rutas de sesión. El navegador la envía automáticamente a
        `/usuarios/refresh` y `/usuarios/logout`; el JavaScript del cliente NUNCA la lee ni la
        escribe. La emite `login`, la rota `refresh` y la limpia `logout`.
  parameters:
    pagina:
      name: pagina
      in: query
      required: false
      description: Número de página (base 1).
      schema:
        type: integer
        minimum: 1
        default: 1
    porPagina:
      name: porPagina
      in: query
      required: false
      description: Cantidad de elementos por página.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    idIdea:
      name: id
      in: path
      required: true
      description: Identificador de la idea.
      schema:
        type: string
        format: uuid
    filtroEstadoIdea:
      name: estado
      in: query
      required: false
      description: Filtra el listado por estado de validación de la idea.
      schema:
        $ref: '#/components/schemas/EstadoIdea'
...
    idGuion:
      name: idGuion
      in: path
      required: true
      description: Identificador del guión de entrevista.
      schema:
        type: string
        format: uuid
...
  schemas:
    CodigoError:
      type: string
      description: |
        Catálogo estable de códigos de error de ValidaLab. El cliente puede ramificar su lógica
        sobre estos valores; el texto de `mensaje` es informativo y puede cambiar.
      enum:
        - VALIDACION_FALLIDA        # 400 — payload inválido o reglas de negocio no satisfechas
        - NO_AUTENTICADO            # 401 — falta token o es inválido/expirado
        - ACCESO_DENEGADO           # 403 — el recurso pertenece a otro owner_id (aislamiento)
        - RECURSO_NO_ENCONTRADO     # 404 — el recurso no existe o no es visible para el usuario
        - CONFLICTO                 # 409 — conflicto de estado (p. ej. transición de outreach inválida)
        - ENTREVISTA_SIN_VINCULO    # 422 — entrevista sin idea o contacto válidos del mismo usuario (RNF-14)
        - API_KEY_INVALIDA          # 422 — la API key BYOK no validó contra el proveedor (RF-20)
        - SALIDA_AGENTE_INVALIDA    # 502 — la salida del agente no cumplió su esquema Zod tras reintentos (RF-AG-03)
        - PROVEEDOR_IA_NO_DISPONIBLE # 503 — el proveedor de IA elegido no respondió (RNF-08)
        - LIMITE_TASA               # 429 — demasiadas solicitudes
        - ERROR_INTERNO             # 500 — error inesperado del servidor
    Error:
      type: object
      description: Sobre estándar de error de toda la API.
      required: [codigo, mensaje]
      properties:
        codigo:
          $ref: '#/components/schemas/CodigoError'
        mensaje:
          type: string
          description: Descripción legible del error (informativa, no estable).
        detalles:
          type: array
          description: Errores de validación campo a campo, cuando aplica.
          items:
            type: object
            properties:
              campo:
                type: string
              problema:
                type: string
    Paginacion:
      type: object
      description: Metadatos de paginación que acompañan a toda colección.
      required: [pagina, porPagina, total, totalPaginas]
      properties:
        pagina:
          type: integer
          description: Página actual (base 1).
        porPagina:
          type: integer
          description: Elementos por página solicitados.
        total:
          type: integer
          description: Total de elementos disponibles para el usuario autenticado.
        totalPaginas:
          type: integer
          description: Total de páginas según `total` y `porPagina`.
    RespuestaPaginada:
      type: object
      description: |
        Sobre de respuesta para colecciones. Los endpoints concretos reemplazan/estrechan
        `datos` con el tipo de recurso correspondiente.
      required: [datos, paginacion]
      properties:
        datos:
          type: array
          items: {}
        paginacion:
          $ref: '#/components/schemas/Paginacion'

    # --- Módulo usuarios (épica E0) ---
...
    EstadoIdea:
      type: string
      description: Estado de validación de una idea.
      enum:
        - borrador        # recién creada; aún sin trabajo de validación
        - en_validacion   # con descubrimiento en curso (se origina con entrevistas, E4)
        - go              # veredicto aprobado: continuar (E6)
        - pivote          # veredicto aprobado: pivotar (E6)
        - kill            # veredicto aprobado: descartar (E6)
        - archivada       # retirada del tablero activo, conservando su evidencia
    Idea:
      type: object
      description: Idea de software en validación. Entidad raíz del dominio, aislada por usuario.
      required: [id, ownerId, titulo, problema, estado, fechaCreacion, fechaActualizacion]
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
          readOnly: true
          description: Propietario de la idea; derivado del token, nunca aceptado como entrada.
        titulo:
          type: string
        descripcion:
          type: string
        problema:
          type: string
        segmentoBeachhead:
          type: string
          description: Segmento de mercado inicial sobre el que se concentra la validación.
        estado:
          $ref: '#/components/schemas/EstadoIdea'
        fechaCreacion:
          type: string
          format: date-time
        fechaActualizacion:
          type: string
          format: date-time
    CrearIdeaRequest:
      type: object
      required: [titulo, problema]
      properties:
        titulo:
          type: string
          minLength: 1
        descripcion:
          type: string
        problema:
          type: string
          minLength: 1
        segmentoBeachhead:
          type: string
    ActualizarIdeaRequest:
      type: object
      description: Cambios de contenido de una idea. No incluye `estado` ni `ownerId`.
      properties:
        titulo:
          type: string
          minLength: 1
        descripcion:
          type: string
        problema:
          type: string
          minLength: 1
        segmentoBeachhead:
          type: string
    IdeasPaginadas:
      description: Página de ideas (sobre `RespuestaPaginada` con `datos` de `Idea`).
      allOf:
        - $ref: '#/components/schemas/RespuestaPaginada'
        - type: object
          properties:
            datos:
              type: array
              items:
                $ref: '#/components/schemas/Idea'

    # --- Hipótesis por idea (épica E2) ---
...
    Pregunta:
      type: object
      description: Pregunta ordenada de un guión de entrevista.
      required: [id, orden, texto]
      properties:
        id:
          type: string
          format: uuid
        orden:
          type: integer
          minimum: 1
          description: Posición de la pregunta dentro del guión (base 1).
        texto:
          type: string
    PreguntaRequest:
      type: object
      description: Pregunta a registrar dentro de un guión.
      required: [texto, orden]
      properties:
        orden:
          type: integer
          minimum: 1
        texto:
          type: string
          minLength: 1
    Guion:
      type: object
      description: Guión de entrevista reutilizable entre ideas, propiedad del usuario.
      required: [id, ownerId, nombre, preguntas, fechaCreacion, fechaActualizacion]
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
          readOnly: true
          description: Propietario del guión; derivado del token, nunca aceptado como entrada.
        nombre:
          type: string
        descripcion:
          type: string
        preguntas:
          type: array
          description: Preguntas del guión en orden.
          items:
            $ref: '#/components/schemas/Pregunta'
        fechaCreacion:
          type: string
          format: date-time
        fechaActualizacion:
          type: string
          format: date-time
    CrearGuionRequest:
      type: object
      description: Alta de guión. No incluye `ownerId` (lo da el token).
      required: [nombre, preguntas]
      properties:
        nombre:
          type: string
          minLength: 1
        descripcion:
          type: string
        preguntas:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PreguntaRequest'
    ActualizarGuionRequest:
      type: object
      description: Cambios sobre un guión (todos opcionales). `preguntas` reemplaza el conjunto ordenado.
      properties:
        nombre:
          type: string
          minLength: 1
        descripcion:
          type: string
        preguntas:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PreguntaRequest'
    GuionesPaginados:
      description: Página de guiones (sobre `RespuestaPaginada` con `datos` de `Guion`).
      allOf:
        - $ref: '#/components/schemas/RespuestaPaginada'
        - type: object
          properties:
            datos:
              type: array
              items:
                $ref: '#/components/schemas/Guion'

    # --- Entrevista y scoring por IA (épica E4) ---
...
  responses:
    NoAutenticado:
      description: Falta el token o es inválido/expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccesoDenegado:
      description: El recurso pertenece a otro `owner_id` (regla de aislamiento multi-tenant).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoEncontrado:
      description: El recurso no existe o no es visible para el usuario autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidacionFallida:
      description: El payload no cumple las reglas de validación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflicto:
      description: Conflicto de estado (p. ej. un `email` ya registrado).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...